Cherry pick changes from commit #8a8d9de to this commit
